+++ Capstone.Application/Services/OrderPromotionsService/OrderPromotionDto.cs
@@ -3,8 +3,8 @@ namespace Capstone.Application.Services.OrderPromotionsService;
 public class OrderPromotionDto
 {
     public Guid Id { get; set; }
-    public decimal MinValue { get; set; }
+    public double MinValue { get; set; }
     public string DiscountType { get; set; } = string.Empty;
-    public decimal DiscountValue { get; set; }
-    public decimal MaxDiscount { get; set; }
+    public double DiscountValue { get; set; }
+    public double MaxDiscount { get; set; }
 }
