+++ Capstone.Domain/Entities/OrderPromotion.cs
@@ -4,10 +4,10 @@ namespace Capstone.Domain.Entities
     {
         public Guid Id { get; set; }
         public Guid PromotionId { get; set; }
-        public decimal MinValue { get; set; }
+        public double MinValue { get; set; }
         public string DiscountType { get; set; } = string.Empty;
-        public decimal DiscountValue { get; set; }
-        public decimal MaxDiscount { get; set; }
+        public double DiscountValue { get; set; }
+        public double MaxDiscount { get; set; }
 
         public Promotion Promotion { get; set; } = null!;
     }
